requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib==1.7.4
pydantic[email]==2.5.0
pydantic-settings==2.1.0
redis==5.0.1
websockets==12.0
aiofiles==23.2.1
pillow==10.1.0
python-magic==0.4.27
email-validator==2.1.0
bleach==6.1.0
cryptography==42.0.4
